@@ -3,78 +3,25 @@
 	import { slide } from 'svelte/transition';
 	import Icon from '@iconify/svelte';
 	import { base } from '$app/paths';
+	import { Dropdown, DropdownItem } from 'flowbite-svelte';
 
-	let committeesHidden = true;
-	let resourcesHidden = true;
+	let isOpen = $state(false);
 
-	$: activeUrl = $page.url.pathname; //this gets the page route which we use to style
+	let activeUrl = $derived($page.url.pathname); //this gets the page route which we use to style
 	const tabsStyle =
 		'font-semibold my-auto transition duration-300 p-1 self-center rounded-md hover:text-slate-900 hover:bg-slate-100';
 	const activeClass =
-		'text-slate-900 bg-slate-100 my-auto font-semibold self-center p-1 rounded-md';
+		'p-1 text-slate-900 bg-slate-100 my-auto font-semibold self-center rounded-md';
 	const dropDownStyle =
-		'w-min p-3 bg-slate-900 space-y-3 text-slate-100 my-auto rounded-md text-center text-nowrap list-none';
+		'p-3 bg-slate-900 space-y-3 text-slate-100 rounded-md text-center';
 </script>
 
 <div
-	class="text-slate-100 bg-slate-900 px-4 fixed top-0 left-0 right-0 flex justify-between z-10 text-sm md:text-lg"
+	class="text-slate-100 bg-slate-900 px-4 sticky top-0 left-0 right-0 flex justify-around z-10 text-sm md:text-lg"
 >
 	<img src="{base}/images/ACMLogo.png" alt="ACM Logo" class="w-36" />
 	<a href="{base}/" class={activeUrl == `${base}/` ? activeClass : tabsStyle}>Home</a>
 	<a href="{base}/about" class={activeUrl == `${base}/about` ? activeClass : tabsStyle}>About Us</a>
-	<div
-		class="relative my-auto"
-		on:mouseover={() => {
-			committeesHidden = false;
-		}}
-		on:mouseleave={() => {
-			committeesHidden = true;
-		}}
-		on:focus={() => {
-			committeesHidden = !committeesHidden;
-		}}
-		aria-haspopup="true"
-		role="button"
-		tabindex="-1"
-	>
-		<button class="flex {tabsStyle}">
-			Committees
-			<Icon
-				icon="ep:arrow-down-bold"
-				class="m-1 duration-150 self-center {committeesHidden ? '-rotate-0' : '-rotate-180'}"
-			/>
-		</button>
-		{#if !committeesHidden}
-			<div class="absolute -left-14 my-auto" transition:slide={{ duration: 300, axis: 'y' }}>
-				<ul class={dropDownStyle}>
-					<li>
-						<a
-							class={activeUrl == `${base}/committees/comppro` ? activeClass : tabsStyle}
-							href="{base}/committees/comppro">Competitive Programming</a
-						>
-					</li>
-					<li>
-						<a
-							class={activeUrl == `${base}/committees/project` ? activeClass : tabsStyle}
-							href="{base}/committees/project">Project Committee</a
-						>
-					</li>
-					<li>
-						<a
-							class={activeUrl == `${base}/committees/robotics` ? activeClass : tabsStyle}
-							href="{base}/committees/robotics">Robotics Committee</a
-						>
-					</li>
-					<li>
-						<a
-							class={activeUrl == `${base}/committees/website` ? activeClass : tabsStyle}
-							href="{base}/committees/website">Website Committee</a
-						>
-					</li>
-				</ul>
-			</div>
-		{/if}
-	</div>
 	<a
 		href="{base}/meetingtimes"
 		class={activeUrl == `${base}/meetingtimes` ? activeClass : tabsStyle}
@@ -84,46 +31,34 @@
 	<a href="{base}/contactus" class={activeUrl == `${base}/contactus` ? activeClass : tabsStyle}>
 		Contact Us
 	</a>
-	<div
-		class="relative my-auto"
-		on:mouseover={() => {
-			resourcesHidden = false;
+	<button
+		class="flex {tabsStyle}"
+		onclick={() => {
+			isOpen = !isOpen;
 		}}
-		on:mouseleave={() => {
-			resourcesHidden = true;
-		}}
-		on:focus={() => {
-			resourcesHidden = !resourcesHidden;
-		}}
-		aria-haspopup="true"
-		role="button"
-		tabindex="-1"
+		>Resources<Icon
+			icon="ep:arrow-down-bold"
+			class="m-1 duration-150 self-center {isOpen ? '-rotate-0' : '-rotate-180'}"
+		/></button
 	>
-		<button class="flex {tabsStyle}"
-			>Resources
-			<Icon
-				icon="ep:arrow-down-bold"
-				class="m-1 duration-150 self-center {resourcesHidden ? '-rotate-0' : '-rotate-180'}"
-			/>
-		</button>
-
-		{#if !resourcesHidden}
-			<div class="absolute -left-10">
-				<ul class={dropDownStyle} transition:slide={{ duration: 300, axis: 'y' }}>
-					<li>
-						<a
-							class={activeUrl == `${base}/resources/joining` ? activeClass : tabsStyle}
-							href="{base}/resources/joining">How to Join</a
-						>
-					</li>
-					<li>
-						<a
-							class={activeUrl == `${base}/resources/teams` ? activeClass : tabsStyle}
-							href="{base}/resources/teams">Microsoft Teams</a
-						>
-					</li>
-				</ul>
-			</div>
-		{/if}
-	</div>
+	<Dropdown
+		{activeUrl}
+		simple
+		class={dropDownStyle}
+		transition={slide}
+		transitionParams={{ duration: 300, axis: 'y' }}
+	>
+		<DropdownItem
+			onclick={() => (isOpen = false)}
+			href="{base}/resources/joining"
+			class={activeUrl == `${base}/resources/joining` ? activeClass : tabsStyle}
+			>How to Join</DropdownItem
+		>
+		<DropdownItem
+			onclick={() => (isOpen = false)}
+			href="{base}/resources/teams"
+			class={activeUrl == `${base}/resources/teams` ? activeClass : tabsStyle}
+			>Microsoft Teams</DropdownItem
+		>
+	</Dropdown>
 </div>
