@@ -1,46 +1,65 @@
 <script>
-	import { Heading, Hr, List, Li } from 'flowbite-svelte';
+	import { Hr, List, Li } from 'flowbite-svelte';
 	import { base } from '$app/paths';
 </script>
 
-<div class="space-y-4 mx-12 md:mx-24 lg:mx-36">
-	<Heading tag="h1" class="mb-4" customSize="text-4xl font-extrabold md:text-5xl lg:text-5xl"
-		>How to Join</Heading
-	>
-	<Hr />
-	<p>
-		Being an active member of the ACM offers many benefits, ranging from access to valuable
-		resources to opportunities for growth. Your participation not only enriches your experience but
-		also helps our chapter give resources and expeience that you can use. We are diligently working
-		towards inproving all student's education and experience, and your involvement is instrumental
-		in this endeavor. <span>Membership dues are $20 for one school year.</span> They are not required
-		but it does help the club afford T-shirts and equipment.
-	</p>
-
-	<div class="space-y-4 2xl:flex 2xl:justify-between">
-		<div class="2xl:w-1/2">
-			<p class="text-xl font-bold text-gray-900 sm:text-xl">Steps to join <span>WSC ACM</span>:</p>
-			<br />
-			<List class="list-decimal space-y-4">
-				<Li>
-					Create an account with <a href="https://accounts.acm.org/">myACM</a>: It’s free and gives
-					you access to <span>ACM</span> resources and opportunities.
-				</Li>
-				<Li>
-					Send your <span>ACM ID</span> to the <span>President</span> or
-					<span>Vice President</span>: This allows us to add you to the <span>WSC ACM</span> chapter,
-					to continue as a club on campus we need at least 20 registered members.
-				</Li>
-				<Li>
-					Attend our <span>ACM</span> meetings: Check the
-					<a href="{base}/meetingtimes">Meeting Times</a>
-					page for more information. If our meeting times don't work with you, you can still participate
-					in <span>ACM</span> with our subcommittees.
-				</Li>
-			</List>
+<div class="h-full w-full py-12 px-6 md:px-12 lg:px-24 xl:px-36 flex items-center text-white">
+	<div class="w-full max-w-7xl mx-auto space-y-12">
+		<!-- Title -->
+		<div class="text-center space-y-4">
+			<p class="text-3xl font-extrabold md:text-4xl">How to Join</p>
+			<Hr />
 		</div>
-		<div>
-			<img src="{base}/images/HowToJoinACM.gif" alt="How to Join Gif" height="100%" width="700" />
+
+		<!-- Description -->
+		<p class="text-lg leading-relaxed text-gray-400">
+			Being an active member of the ACM offers many benefits, ranging from access to valuable
+			resources to opportunities for growth. Your participation not only enriches your experience
+			but also helps our chapter provide tools and events that support your success. We are
+			diligently working to improve student education and experience—and your involvement is
+			instrumental in this.
+			<span class="font-bold">Membership dues are $20 for one school year.</span> These are optional
+			but help us cover T-shirts, equipment, and snacks.
+		</p>
+
+		<!-- Steps + Image -->
+		<div class="flex flex-col-reverse xl:flex-row xl:items-center xl:justify-between gap-12">
+			<!-- Steps -->
+			<div class="w-full xl:w-1/2 space-y-6">
+				<p class="text-xl font-bold text-white">
+					Steps to join <span class="font-bold">WSC ACM</span>:
+				</p>
+
+				<List class="list-decimal list-inside space-y-4 text-gray-400">
+					<Li>
+						Create an account at
+						<a href="https://accounts.acm.org/" target="_blank" class="text-blue-600 underline"
+							>myACM</a
+						>. It’s free and gives you access to <span class="font-bold">ACM</span> resources and opportunities.
+					</Li>
+					<Li>
+						Send your <span class="font-bold">ACM ID</span> to the
+						<span class="font-bold">President</span>
+						or
+						<span class="font-bold">Vice President</span> so we can officially add you to the chapter
+						roster.
+					</Li>
+					<Li>
+						Start attending meetings! See our
+						<a href="{base}/meetingtimes" class="text-blue-600 underline">Meeting Times</a>. If you
+						can't make the time, you can still participate in our subcommittees and events.
+					</Li>
+				</List>
+			</div>
+
+			<!-- Image -->
+			<div class="w-full xl:w-1/2 flex justify-center">
+				<img
+					src="{base}/images/HowToJoinACM.gif"
+					alt="How to Join ACM"
+					class="rounded-xl shadow-lg max-w-full h-auto"
+				/>
+			</div>
 		</div>
 	</div>
 </div>
@@ -48,9 +67,5 @@
 <style>
 	span {
 		font-weight: bold;
-	}
-
-	a {
-		color: blue;
 	}
 </style>
